feat(link): add utils file
